+++ packages/json-rpc/polkadot-sdk-compat/src/parsed-enhancers/index.ts
@@ -1,4 +1,5 @@
 export { fixChainSpec } from "./chain-spec"
+export { fixDescendantValues } from "./fix-descendant-values"
 export { fixUnorderedBlocks } from "./fix-unordered-blocks"
 export { fixUnorderedEvents } from "./fix-unordered-events"
 export { patchChainHeadEvents } from "./patch-chainhead-events"