+++ packages/json-rpc/polkadot-sdk-compat/CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Added
 
 - Add `fixChainSpec` enhancer which addresses [the following issue](https://github.com/paritytech/polkadot-sdk/issues/5539) on the PolkadotSDK node.
+- Add `fixDescendantValues` enhancer which addresses [the following issue](https://github.com/paritytech/polkadot-sdk/issues/5589) on the PolkadotSDK node.
 
 ## 2.1.0 - 2024-09-04